+++ apricot-backend/src/main/java/com/apricot/core/business/repository/company/CompanyRepository.java
@@ -16,5 +16,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CompanyRepository extends JpaRepository<Company,Long> {
-    List<Company> findByUser_Id(Integer id);
+    List<Company> findByUser_Id(Long id);
 }
